[507431] Move IO lib and add Handle descriptors
Moves the io implementation from the fblib to the core folder. The
embrick implementation is adapted accordingly. Additionally the
BitSlaveHandle is introduced as global lib as it is used in many IO
implementations.

Handle descriptors are introduced which are used to exchange information
about handles between the device controller and the corresponding
configuration fb. The IO implementation should extend the Handle struct
with properties, which uniquely identify the handle.

+++ src/modules/embrick/slave/packages.h
@@ -40,8 +40,8 @@ struct SlaveInit {
 
     pkg.deviceId = ntohs(pkg.deviceId);
     // Switch bytes of deviceId as it is transmitted with a different endianess
-    pkg.deviceId = ((pkg.deviceId & 0xFF00) >> 8)
-        | ((pkg.deviceId & 0xFF) << 8);
+    pkg.deviceId = (uint16_t) (((pkg.deviceId & 0xFF00) >> 8)
+        | ((pkg.deviceId & 0xFF) << 8));
 
     pkg.producerId = ntohs(pkg.producerId);
